fe: reorganized navbar folder
