+++ frontend/src/lib/partials/global/navbar/navbarDivider.svelte
