+++ frontend/src/lib/partials/global/navbar/navbarMenuLink.svelte
