+++ frontend/src/routes/__layout.svelte
@@ -8,7 +8,7 @@
 	import { req } from '$lib/requestUtils';
 
 	import { Loading } from '$lib/components';
-	import Navbar from '$lib/partials/navbar.svelte';
+	import Navbar from '$lib/partials/global/navbar/navbar.svelte';
 
 	//
